% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_ve_mock_data.R
\docType{data}
\name{ve_mock_data}
\alias{ve_mock_data}
\title{ve_mock_data}
\format{
A data frame with 52 rows and 2 variables:
\describe{
\item{ week }{  Date }
\item{ ve }{  numeric: vaccine effectiveness }
}
}
\source{
Mock data for vaccine effectiveness
}
\usage{
ve_mock_data
}
\description{
Vaccine effectiveness data.
}
\keyword{datasets}
