% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_coverage_and_incidence_mock_data.R
\docType{data}
\name{coverage_and_incidence_mock_data}
\alias{coverage_and_incidence_mock_data}
\title{coverage_and_incidence_mock_data}
\format{
A list with two data frames:
\describe{
\item{ incidence_data }{  data.frame with weekly incidence data }
\item{ coverage_data }{  data.frame with weekly coverage data }
}
}
\source{
Simulated coverage and incidence data
}
\usage{
coverage_and_incidence_mock_data
}
\description{
Coverage and incidence mock data. Coverage values are computed considering a sample size of 1234 individuals.
}
\keyword{datasets}
