% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_coverage_and_incidence_mock_data.R
\docType{data}
\name{coverage_and_incidence_mock_data}
\alias{coverage_and_incidence_mock_data}
\title{coverage_and_incidence_mock_data}
\format{
A data frame with  rows and  variables:
\describe{
\item{ incidence_data }{  data.frame }
\item{ coverage_data }{  data.frame }
}
}
\source{
Source
}
\usage{
coverage_and_incidence_mock_data
}
\description{
Description.
}
\keyword{datasets}
