% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events_avertable_by_increasing_coverage.R,
%   R/events_averted_by_vaccination.R, R/number_needed_to_vaccinate.R
\name{compute_events_avertable_by_increasing_coverage}
\alias{compute_events_avertable_by_increasing_coverage}
\alias{compute_events_averted_by_vaccination}
\alias{compute_number_needed_to_vaccinate_machado}
\alias{compute_number_needed_to_vaccinate_tuite_fisman}
\title{Compute events averted by increasing the final vaccine coverage}
\usage{
compute_events_avertable_by_increasing_coverage(
  number_of_events,
  cumulative_coverage,
  vaccine_coverage_increase,
  vaccine_effectiveness
)

compute_events_averted_by_vaccination(
  number_of_events,
  cumulative_coverage,
  vaccine_effectiveness
)

compute_number_needed_to_vaccinate_machado(
  number_of_events,
  number_of_events_averted,
  pop_at_risk,
  vaccine_effectiveness
)

compute_number_needed_to_vaccinate_tuite_fisman(
  number_of_vaccinated,
  number_of_events_averted
)
}
\arguments{
\item{number_of_events}{number of events}

\item{cumulative_coverage}{cumulative vaccination coverage}

\item{vaccine_coverage_increase}{percentage increase in final vaccine coverage (between 0 and 1)}

\item{vaccine_effectiveness}{vaccine effectiveness}

\item{number_of_events_averted}{number of events averted}

\item{pop_at_risk}{population at risk}

\item{number_of_vaccinated}{number of vaccinated individuals}
}
\value{
a list with the new vaccine coverage ("new_vaccine_coverage") and the estimated number of events averted ("nabe")

estimated number of events averted

The number of individuals needed to vaccinate to avert
one event

The number of individuals needed to vaccinate to avert
one event
}
\description{
Compute events averted by increasing the final vaccine coverage

Compute events averted by vaccination

Compute the number of individuals needed to vaccinate to prevent
one event according to Machado et al. method

Compute the number of individuals needed to vaccinate to prevent
one event according to Tuite and Fisman method
}
\details{
The number of events averted by vaccination is calculated as described by Machado et al. (2019) \url{doi:10.2807/1560-7917.ES.2019.24.45.1900268}.

The number of individuals needed to vaccinate to prevent one event is calculated as described by Machado et al. (2019) \url{doi:10.2807/1560-7917.ES.2019.24.45.1900268}.

The number of individuals needed to vaccinate to prevent one event is calculated as described by Tuite and Fisman (2013) \url{doi:10.1016/j.vaccine.2012.11.097}.
}
